+++ CHANGELOG.rst
@@ -3,8 +3,8 @@ Changelog
 This log contains a list of bug fixes and added, modified or even removed
 features.
 
-0.2 - [Codename to be selected]
--------------------------------
+0.2 - Boggart
+-------------
 *Currently in development*
 
 - Added :class:`brownie.itools.chain`.